Refactored code + Fixed bugs
Changed many of the existing functions to be more customisable when called

Fixed the URL generation - many  parameters can be changed now

Fixed a bug with ID generation

Review cursor variable now has use and allows more than the maximum of 100 reviews to be grabbed - This feature has been tested although it's not currently in this version

Added some more descriptive comments to the code

Changed the method of storing the data to the final list

Current state of this program only needs to properly grab the total number of reviews
If this number is more than 5000, the total number that will be grabbed will be 5000 (Question to ask may be if more than 5000 want to be grabbed, and stored in a different file, wording in the brief suggests this could be an option)
If the number is less than 5000, divide it by 100 to get the least number of total full 100 requests. the decimal point number will always be the leftover number to grab.
e.g 3467/100 = 34.67
Grab 100 34 times
0.67 is leftover
67 reviews are remaining to grab